ai_service/requirements.txt
fastapi
uvicorn
openai>=1.0.0
numpy
pydantic
python-dotenv
